rack: rename «parsers» folder to «lib»
