+++ rack/lib/Iniparse/spec/test_windows.ini
