+++ rack/state/fx.lua
@@ -2,7 +2,7 @@
 This class keeps track of an FX’s internal state (list of parameters, custom layouts, etc.)
 Each FX represented in the parent `state` gets its own instance of this class, instantiated with `fx.new()`.
 ]]
-local IniParse = require("parsers.Iniparse.IniParse")
+local IniParse = require("lib.Iniparse.IniParse")
 local table_helpers = require("helpers.table")
 local defaults = require("helpers.defaults")
 local layout_enums = require("state.layout_enums")
